+++ src/bpp/models/wydawnictwo_ciagle.py
@@ -79,6 +79,13 @@ class Wydawnictwo_Ciagle_Autor(
         ]
 
     def pbn_get_json(self):
+        if (
+            not self.afiliuje
+            or not self.jednostka.skupia_pracownikow
+            or self.jednostka.pk < 0
+        ):
+            return
+
         ret = {
             # to NIE jest ta flaga; to jest flaga dot. czy dane są w indeksie ORCID
             # a nie czy autor ma Orcid ID
@@ -99,9 +106,18 @@ class Wydawnictwo_Ciagle_Autor(
         #    ret["institutionId"] = self.jednostka.pbn_uid.pk
 
         if self.autor.pbn_uid_id:
-            ret["personObjectId"] = self.autor.pbn_uid.pk
+            scientist = self.autor.pbn_uid
+
+            pesel = scientist.value(
+                "object", "externalIdentifiers", "PESEL", return_none=True
+            )
+            if pesel:
+                ret["personNaturalId"] = pesel
+
+            ret["personObjectId"] = scientist.pk
+
         else:
-            if self.autor.orcid:
+            if self.autor.orcid and self.autor.orcid_w_pbn:
                 ret["personOrcidId"] = self.autor.orcid
 
         if not ret.get("personOrcidId") and not ret.get("personObjectId"):
@@ -302,11 +318,11 @@ class Wydawnictwo_Ciagle(
                     author["affiliations"] = [jednostka.pbn_uid_id]
                     jednostki.add(elem.jednostka)
 
-            authors.append(author)
+                    statement = elem.pbn_get_json()
+                    if statement:
+                        statements.append(statement)
 
-            statement = elem.pbn_get_json()
-            if statement:
-                statements.append(statement)
+            authors.append(author)
 
         ret["authors"] = authors
         ret["statements"] = statements