+++ HISTORY.rst
@@ -3,6 +3,16 @@
 Historia zmian
 ==============
 
+202105.65
+---------
+
+* eksportuj naturalId w danych z PBN (#1063),
+* lepsze matchowanie źródeł z PBN (#1064),
+* weryfikuj obecnośc ORCID w PBN dla niezmatchowanych autorów (#1054),
+* pobieraj wszystkie osoby z PBNu (b/n),
+* pole dla wpisania wartości, czy praca występuje w profilu ORCID autora (#1054),
+* nie eksportuj oświadczeń dla autorów bez afiliacji (#1055),
+
 202105.64
 ---------